mini_printf: Added glibc-like NULL termination.
Originally, mini_printf does not NULL terminate the output.
Henceforth mini_vsnprintf will always append a NULL character,
as long as the buffer size is non-zero, like glibc does.

Keep in mind that the output will be truncated in order to make
space for the terminator, if there's not enough space for both.